requirements.txt
# Microsoft Agent Framework (preview - requires --pre flag)
agent-framework-azure-ai --pre

# Environment and utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Email handling
aiosmtplib>=3.0.0
aioimaplib>=1.0.0

# HTTP client for APIs
httpx>=0.27.0
aiohttp>=3.9.0

# LinkedIn API client
requests-oauthlib>=1.3.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
